final nits. Move simpleDurationTimeout back to package private
